TG: main.m - actually use full double precision for frametime; ODScene - likewise; ODPEntity - likewise; fix fallout following this changes

+++ AppsV2/OceanDemo/Entities/ODBasePlane.h
@@ -27,7 +27,7 @@
 
 - (void) setCamera:(ODCamera *)newCamera;
 
-- (void) update:(const float)frameTime;
+- (void) update:(const double)frameTime;
 - (void) render;
 
 @end
